Se crea content-analysis-controller, encargado de manejar las rutas relacionadas al analisis de contenido
